+++ arenera.jl
@@ -9,11 +9,13 @@ pa_type = jOpqua.PathogenType(
     "pa_type",
     10,
     "ARNDCEQGHILKMFPSTWYV*",
+    0,
+    1.0,
     SA[ # pathogen_coefficient_functions::SVector{NUM_COEFFICIENTS,Function} # Each element takes seq argument, returns Float64
         g->1.0, g->1.0, g->1.0, g->1.0, g->1.0,
         g->1.0, g->1.0, g->1.0, g->1.0, g->1.0,
         g->1.0, g->1.0, g->1.0, g->1.0, g->1.0,
-        g->1.0,
+        g->1.0, g->1.0,
     ],
 )
 
@@ -23,13 +25,13 @@ re_type = jOpqua.ResponseType(
         (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
         (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
         (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
-        (imp_g, mat_g)->1.0,
+        (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
     ],
     SA[ # specific_coefficient_functions::SVector{NUM_COEFFICIENTS,Function},
         (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
         (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
         (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
-        (imp_g, mat_g, pat_g)->1.0,
+        (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
     ],
 )
 
@@ -38,8 +40,8 @@ class_parameters = jOpqua.ClassParameters(
     SA[ # base_coefficients::SVector{NUM_COEFFICIENTS,Float64}
         0.0, 1.0, 0.0, 0.0, 1.0,
         0.0, 0.0, 0.0, 0.0, 0.0,
-        0.0, 0.0, 0.0, 0.0, 1.0,
-        0.0
+        0.0, 0.0, 0.0, 0.0, 0.0,
+        1.0, 0.0,
     ],
     Dict(), # class_change_fractions::Dict{String,Float64} # size CLASSES, must sum to 1
     Dict(), # inter_population_contact_fractions::Dict{String,Float64} # size POPULATIONS, must sum to 1
@@ -57,7 +59,7 @@ pop = jOpqua.newPopulation!("pop", model)
 class = jOpqua.newClass!("class", class_parameters, pop)
 host = jOpqua.newHost!(class, pop, model)
 pat = jOpqua.newPathogen!("ATCG", class, pa_type)
-res = jOpqua.newResponse!(pat, pat, (0,0), class, re_type)
+res = jOpqua.newResponse!(pat, pat, (pat.sequence,pat.sequence,re_type.id), class, re_type)
 
 jOpqua.addPathogenToHost!(pat, 1, class, pop, model)
 jOpqua.addResponseToHost!(res, 1, class, pop, model)
@@ -65,9 +67,26 @@ jOpqua.addResponseToHost!(res, 1, class, pop, model)
 jOpqua.removePathogenFromHost!(1, 1, class, pop, model)
 jOpqua.removeResponseFromHost!(1, 1, class, pop, model)
 
-jOpqua.choosePathogen(1, 1, 1, 1, model, rand())
-jOpqua.chooseResponse(1, 1, 1, 7, model, rand())
-jOpqua.chooseHost(1, 1, 7, model, rand())
-jOpqua.chooseClass(1, 7, model, rand())
-jOpqua.choosePopulation(7, model, rand())
-jOpqua.chooseEvent(model, rand())
+rand_n = rand()
+x, rand_n = jOpqua.choosePathogen(1, 1, 1, 1, model, rand_n)
+x, rand_n = jOpqua.chooseResponse(1, 1, 1, 7, model, rand_n)
+x, rand_n = jOpqua.chooseHost(1, 1, 7, model, rand_n)
+x, rand_n = jOpqua.chooseClass(1, 7, model, rand_n)
+x, rand_n = jOpqua.choosePopulation(7, model, rand_n)
+x, rand_n = jOpqua.chooseEvent(model, rand_n)
+
+jOpqua.addPathogenToHost!(pat, 1, class, pop, model)
+jOpqua.addResponseToHost!(res, 1, class, pop, model)
+
+x, rand_n = jOpqua.choosePathogen(1, 1, 1, 1, model, rand_n)
+x, rand_n = jOpqua.chooseResponse(1, 1, 1, 7, model, rand_n)
+x, rand_n = jOpqua.chooseHost(1, 1, 7, model, rand_n)
+x, rand_n = jOpqua.chooseClass(1, 7, model, rand_n)
+x, rand_n = jOpqua.choosePopulation(7, model, rand_n)
+x, rand_n = jOpqua.chooseEvent(model, rand_n)
+
+jOpqua.establishMutant!(model, rand())
+jOpqua.clearPathogen!(model, rand())
+jOpqua.acquireResponse!(model, rand())
+jOpqua.establishRecombinant!(model, rand())
+jOpqua.intraPopulationContact!(model, rand())