+++ src/simulation/constants.jl
@@ -6,7 +6,7 @@ const MUTANT_ESTABLISHMENT = 1
 const CLEARANCE = 2
 const RESPONSE_ACQUISITION = 3
 # Two pathogens, single host
-const RECOMBINATION = 4
+const RECOMBINANT_ESTABLISHMENT = 4
 # Single pathogen, two hosts
 const INTRA_POPULATION_CONTACT = 5
 const INTER_POPULATION_CONTACT = 6
@@ -26,20 +26,21 @@ const MIGRATION = 11
 
 # Choice modifiers, zooming in in scale (order matters)
 # Event choice
-const MUTATION_AT_CONTACT = 12
+const RECOMBINATION_PER_REPLICATION = 12
+const MUTATION_PER_REPLICATION = 13
 # Population choice
-const RECEIVE_MIGRATION = 13
+const RECEIVE_MIGRATION = 14
 # Class choice
-const RECEIVE_CLASS_CHANGE = 14
+const RECEIVE_CLASS_CHANGE = 15
 # Host choice
-const RECEIVE_CONTACT = 15
+const RECEIVE_CONTACT = 16
 # Pathogen choice
-const INTRAHOST_FITNESS = 16
+const INTRAHOST_FITNESS = 17
 
 # Global trackers
 const PATHOGEN_EVENTS = SA[
     MUTANT_ESTABLISHMENT, CLEARANCE, RESPONSE_ACQUISITION,
-    RECOMBINATION, INTRA_POPULATION_CONTACT, INTER_POPULATION_CONTACT
+    RECOMBINANT_ESTABLISHMENT, INTRA_POPULATION_CONTACT, INTER_POPULATION_CONTACT
 ]
 const NUM_PATHOGEN_EVENTS = length(PATHOGEN_EVENTS)
 
@@ -53,7 +54,7 @@ const EVENTS = SA[PATHOGEN_EVENTS..., RESPONSE_EVENTS..., HOST_EVENTS...]
 const NUM_EVENTS = length(EVENTS)
 
 const CHOICE_MODIFIERS = SA[
-    MUTATION_AT_CONTACT, RECEIVE_MIGRATION,
+    RECOMBINATION_PER_REPLICATION, MUTATION_PER_REPLICATION, RECEIVE_MIGRATION,
     RECEIVE_CLASS_CHANGE, RECEIVE_CONTACT, INTRAHOST_FITNESS
 ]
 const NUM_CHOICE_MODIFIERS = length(CHOICE_MODIFIERS)
@@ -63,7 +64,10 @@ const NUM_COEFFICIENTS = length(COEFFICIENTS)
 
 # Starter coefficients
 const START_COEFFICIENTS = SVector{NUM_COEFFICIENTS,Float64}(
-    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0]
+    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0]
 # class change, inter-population contact and migration numbers per class or
 # population are fractions that sum to one, so no need to account for in here
 )
+
+# Misc constants
+const CHROMOSOME_SEPARATOR = "/"