+++ src/simulation/initializers.jl
@@ -1,25 +1,23 @@
 using StaticArrays
 
 function newPathogen!(sequence::String, class::Class, type::PathogenType)
-    class.pathogens_idx += 1
-    class.pathogens[class.pathogens_idx] = Pathogen(
-        class.pathogens_idx, sequence, pathogenSequenceCoefficients(sequence, type), type
+    class.pathogens[sequence] = Pathogen(
+        sequence, pathogenSequenceCoefficients(sequence, type), type
     )
 
-    return class.pathogens[class.pathogens_idx]
+    return class.pathogens[sequence]
 end
 
 function newResponse!(
-    imprinted_pathogen::Pathogen, matured_pathogen::Pathogen, parent::Tuple{Int64,Int64},
+    imprinted_pathogen::Pathogen, matured_pathogen::Pathogen, parent::Tuple{String,String,String},
     class::Class, type::ResponseType)
-    class.responses[(imprinted_pathogen.id,matured_pathogen.id)] = Response(
-        (imprinted_pathogen.id,matured_pathogen.id), parent::Tuple{Int64,Int64},
-        imprinted_pathogen, matured_pathogen,
+    class.responses[(imprinted_pathogen.sequence, matured_pathogen.sequence, type.id)] = Response(
+        parent, imprinted_pathogen, matured_pathogen,
         responseStaticCoefficients(imprinted_pathogen.sequence, matured_pathogen.sequence, type),
         type
     )
 
-    return class.responses[(imprinted_pathogen.id,matured_pathogen.id)]
+    return class.responses[(imprinted_pathogen.sequence,matured_pathogen.sequence,type.id)]
 end
 
 function newHost!(class::Class, population::Population, model::Model)
@@ -46,9 +44,8 @@ end
 function newClass!(id::String, parameters::ClassParameters, population::Population)
     push!(population.classes, Class(
         id, parameters,
-        Dict{Int64,Pathogen}(), Dict{Int64,Response}(),
+        Dict{String,Pathogen}(), Dict{Tuple{String,String,String},Response}(),
         Vector{Host}(undef, 0),
-        0,
         Matrix{Float64}(undef, NUM_EVENTS, 0),
         Matrix{Float64}(undef, NUM_CHOICE_MODIFIERS - 1, 0),
     ))